Introduce env var global double mutex
This introduces a new `build::environment` module with moved `Environment` RAII guard and new global `EnvironmentLock` and related. The lock has to be a double one, because we might need doubly scoped env vars (outer one for Cargo routine and inner one for underlying rustc calls across different threads, contained within Cargo routine invocation scope). While the lock itself and provided types don't enforce a strict lock ordering, the order of retrieved locks and types is still encoded in the type system.
This might not be a complete and fully abstracted away solution to N-valued mutex and guarded environment, but works good enough and seems to cover enough cases in both production and testing scenarios.