Create simple dep-graph from Cargo metadata
Incomplete `build::plan::Plan` is created, containing an inter-package dependency graph and targets/features of packages in the workspace scope, and stored when performing the Cargo build routine. It's not currently used and specific data held will require further design and work.
For now copies some of the Cargo types, as it's not yet publicly exposed.