Append the label value of the primary spans of diagnostics to the LSP diagnostics message if it is available.
This is very useful in diagnostic messages where the label has specific
diagnostics information for the site and little in the message. The most
popular message and an example is detailed as follows:

In cases like E0308 or better known as "mismatched types", the message
is simply "mismatched types".

Users seeing "mismatched types" in their editing buffers would have had
to immediately step out of their buffers and run whatever compiles the
line in their terminals to take a look at what is described here as the
label in order to figure out what and how the types are in conflict.

Related: https://github.com/rust-lang-nursery/rls-vscode/issues/103