Initialize RlsExecutor during `new()` instead of during `init()`
`fn init()` is called each time `rustc` command is prepared, so we ended up calling it as many times as there were `rustc` invocations for the package/workspace.
To initialize the packages we only need `Workspace`, which we do create already when creating `RlsExecutor` - so just pass it and use it when calling `RlsExecutor::new()`.