Use cached build plan under workspace_mode if available to perform build
Because of how cached build plan works, under `workspace_mode` every member
package must be forcefully rebuilt to store the arguments Cargo would generate
for the target. This is required when only a certain member package will be
rebuilt initially, but changes made to another one will pull a rebuild of a
package whose rustc call wasn't cached.
Furthermore, sometimes user can quickly input changes and save, before any build
can be made. When this happens and there are no modified files, build plan will
return an empty job queue, since no files are dirty at that time. When that
happens, it delegates to Cargo, just when build scripts are modified. In the
future, detecting dirty files and guessing which files relate to which rebuilt
package is a concrete area for improvement.