Track versioned dirty files since last build
Previously only `rustc` was called on a single file or whole `cargo` routine
was executed  to perform the build and since the build plan, cached from the
`cargo` run, is used now in the `workspace_mode` instead, RLS needs to feed
it files changed since last build.
Since while the build is being performed, the user can further modify files
that are currently being built, only files that were modified prior to that
will be marked as clean.