Remove intermittent failing test
The `bin_lib_project` and `bin_lib_project_no_cfg` tests can
intermittently fail when run in parallel, especially when running
them in the rustc x.py tests. This commit temporarily removes one of
those tests. The test is still legitimate, but we need to find a way
to be able to use the same test project directory (lib_bin) in tests,
and make sure that they work when testing in the CI environment.