Implement Deglob as a CodeAction
Provide code actions for deglobbing import for wildcard imports, except
for "prelude::*". To avoid false positives in the code actions, the
deglob command is run before the code action is presented to the user.

The original custom LSP extension is still in place in works. As such
the `Deglob::handle` function is split into a new
`calculate_workspace_edit`, which calculates the data for the deglob
action and is not used to the code action. The original `Deglob::handle`
just calls this function and is reponsible for sending the messages to
the client.

Addresses https://github.com/rust-lang-nursery/rls/issues/515