Document the `actions` module
This makes the `actions` module "public" so that the `#![deny(missing_docs)]`
pragma applies to it, and adds documentation for all of its `pub` items.

I put "public" in quotes because this crate is a `[[bin]]` and not a `[lib]` so
nothing external can depend on it, and since it is a top level module every
other module in the crate can already use it.