Document the `config` module
This commit makes the `config` module "public" so that the
`#![deny(missing_docs)]` pragma applies to it. I don't know the details of each
of `Config`'s fields, so I opted to allow missing documentation there for
now. Hopefully someone who does know can fill it in and remove that exception ;)