Include secondary spans as LSP related information
Also includes line numbers to every code suggestion that's related
to a span outside of the primary one.

However, this approach does not yet fully work with diagnostics like
E0495. These include children with single primary span, but with
each children there's only a meaningful message - label is not defined.
Because we flatten all the spans and only use them if the label is defined,
these children messages are ignored for now.