Introduce Sender trait and use it instead of Output.
Currently Output is used for two separate things - to send responses to
incoming requests and to send new requests and notifications. This
changelist is a first step to separate those two responsibilities, and
make it clear what objects actually use Output just for sending new
requests. It also nicely hides the need to handle the request id.

Also, Output currently only writes data to client, which means that
`Output.request` method will never be able to nicely return a response
from client. The plan with Sender is to nicely handle this case long
term.