Filter out duplicate symbols
Sometimes analysis returns duplicate symbols for the same location. It
can be a bug in the analysis, however in general it might be the case
that two logically different symbols correspond to the same location,
if we use `include!` or something like that.

So, let's do duplicate detection before submitting results to the user
for the time being