Enhanced hover tooltips
Hover tooltips now extract documentation and declarations from
source based on the the file and span information that comes out
of the save-analysis. If the save-analysis data is missing (i.e.
there is no `Def` available), racer is used as fallback and the
`Def` is sythesized from the racer `Match` if one is found.

Function and object declarations are formatted using rustfmt and
context for local variables is also included in the tooltips.

All documentation that is discovered is now post-processed in order
to clean up code blocks so that syntax coloring works in in editors
that support markdown.

The racer fallback can be disabled via the `racer_completion`
configuration flag.