Support external builds
This is currently enabled via `build_command` config option.
The specified command runs as a build procedure; it should output a
list of save-analysis .json files to be reloaded by the RLS so that the
usual features, e.g. type info on hover, work.

The `build_command` requires `build_on_save` (otherwise RLS needs a lot
of dummy rebuilds to sync the files and treat them as fresh)