Implement external build plan
This adds a `BuildGraph` trait which aims to abstract away the API for
a build plan and it implements it for the in-process Cargo build plan
as well as the new, external one (for `cargo build --build-plan`
format).
In addition to that, since save-analysis (since rls-data 0.18.1)
includes the invocation used to compile a given crate, we try to
recreate the external build plan from the passed save-analysis files
via `build_command` config option.