Add serde implementation to complement rustc-serialize
We want to transition off of rustc-serialize and onto serde, so this
commit sets up both to co-exist, gated by feature flags. Due to the fact
that they implement the same functions, only one feature flag can be
active at a time. To compile rls-analysis with serde, it is necessary to
run cargo build with the `--no-default-features` flag and the
`--features "serialize-serde"` flag.

The tests in this commit will still work with the default feature flags
but will need to be changed before the serde implementation can pass
tests.