Added limited support for typing code lenses
For now, it only shows types for variables that are being declared alone with a `let (mut) var`, and refreshes slightly too early (you need to type a bit to wake it up after just writing a new declaration.
It also does no typename compacting, so the fact that I haven't made the lenses inline is a plus for now.