Test typos with autocfg
In `tests/client.rs`, `client_dependency_typo_and_fix` was failing in my
Fedora builds of the rust toolchain. That uses `--enable-vendor`, and
the source package only has `version_check 0.9`, but this test is trying
to correct to `version_check 0.1`.

Nothing against `version_check`, but I ported this test to use `autocfg`
instead. It's also fast to compile and available in the vendored source
package, and it has a stable version 1 so this problem won't recur.