chore: Upgrade cargo
Cargo has switched exclusively to `toml_edit`, so when downcasting
errors to get line and column, we need to downcast for those errors as
well.

This does not attempt to port rls to `toml_edit` and an analysis was not
done to verify if the `find_toml_error` code path only needs to check
for one toml library.

The downside to this solution is its brittle.  Any time cargo upgrades
through a breaking `toml_edit` version, this test will break again and
we'll have to upgrade rls to fix it.