Inline racer's source code into RLS
This deletes a bunch of the irrelevant test infrastructure to minimize the added
size, and is overall a relatively small addition. It'll make it much easier for
us to bump RLS going forward in rust-lang/rust as we fully own the bits that
break due to rustc changes now.

In practice neither racer nor RLS are really under active maintenance so this
doesn't come at a significant ongoing cost.